rename some file paths
