+++ lib/appium_lib/core/common/error.rb
