+++ lib/appium_lib/core/common/wait.rb
